requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
requests
